Keep Mistral import regressions focused
